+++ lib/app/modules/changePassword/bindings/change_password_binding.dart
@@ -1,12 +1,15 @@
 import 'package:get/get.dart';
 
 import '../controllers/change_password_controller.dart';
+import '../states/change_password_state.dart';
 
 class ChangePasswordBinding extends Bindings {
   @override
   void dependencies() {
     Get.lazyPut<ChangePasswordController>(
-      () => ChangePasswordController(),
+      () => ChangePasswordController(
+        state: Get.put(ChangePasswordState()),
+      ),
     );
   }
 }
