+++ lib/app/modules/changePassword/controllers/change_password_controller.dart
@@ -1,9 +1,17 @@
 import 'package:get/get.dart';
+import '../states/change_password_state.dart';
 
 class ChangePasswordController extends GetxController {
-  //TODO: Implement ChangePasswordController
+  ChangePasswordState state;
+
+  ChangePasswordController({required this.state});
+
+  //TODO: Implement NewPasswordController
+
+  void handleClick() {}
 
   final count = 0.obs;
+
   @override
   void onInit() {
     super.onInit();
