+++ lib/app/modules/changePassword/states/change_password_state.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/cupertino.dart';
 import 'package:get/get.dart';
 
-class NewPasswordState extends GetXState {
+class ChangePasswordState extends GetXState {
   TextEditingController oldPasswordController = TextEditingController();
   TextEditingController newPasswordController = TextEditingController();
   TextEditingController confirmPasswordController = TextEditingController();
