+++ lib/app/modules/forgotPassword/bindings/forgot_password_binding.dart
@@ -1,0 +1,12 @@
+import 'package:get/get.dart';
+
+import '../controllers/forgot_password_controller.dart';
+
+class ForgotPasswordBinding extends Bindings {
+  @override
+  void dependencies() {
+    Get.lazyPut<ForgotPasswordController>(
+      () => ForgotPasswordController(),
+    );
+  }
+}
