+++ lib/app/modules/forgotPassword/controllers/forgot_password_controller.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 
-class PasswordViewController extends GetxController {
+class ForgotPasswordController extends GetxController {
   TextEditingController otpController = TextEditingController();
   TextEditingController newPasswordController = TextEditingController();
   TextEditingController confirmPasswordController = TextEditingController();
