--- lib/app/modules/password_view/bindings/password_view_binding.dart
@@ -1,12 +1,0 @@
-import 'package:get/get.dart';
-
-import '../controllers/password_view_controller.dart';
-
-class PasswordViewBinding extends Bindings {
-  @override
-  void dependencies() {
-    Get.lazyPut<PasswordViewController>(
-      () => PasswordViewController(),
-    );
-  }
-}
